+++ packages/plugins/src/utils.ts
@@ -5,7 +5,7 @@ import type { SwapKitPluginParams } from "./types";
 export function createPlugin<
   const Name extends string,
   T extends (params: SwapKitPluginParams) => Record<string, unknown>,
-  K extends { supportedSwapkitProviders?: (ProviderName | string)[] },
+  K extends { supportedSwapkitProviders?: readonly ProviderName[] },
 >({ name, properties, methods }: { name: Name; properties?: K; methods: T }) {
   function plugin(pluginParams: SwapKitPluginParams) {
     return { ...methods(pluginParams), ...properties } as K & ReturnType<T>;